fix: use UA_VT.Int16 for xs:int — Faaster maps xs:int to OPC UA type 4 (Int16)

+++ sources/i_o_interface/cobot_io.py
@@ -284,17 +284,17 @@ class CobotIOInterface(ISubmodelExtension):
         meta = self._ctx.get_node(path)
         if meta:
             try:
-                # Python type → OPC UA VariantType mapping for AAS properties:
-                #   xs:float   → Float   (10)  but Python float is Double (11)
-                #   xs:int     → Int32   (4)   but Python int   is Int64  (8)
-                #   xs:boolean → Boolean (1)   Python bool maps correctly
-                # bool must be checked before int (bool is a subclass of int).
+                # Faaster OPC UA VariantType mapping for AAS valueTypes:
+                #   xs:boolean → Boolean (1)   Python bool inferred correctly
+                #   xs:float   → Float   (10)  Python float inferred as Double (11)
+                #   xs:int     → Int16   (4)   Python int  inferred as Int64  (8)
+                # bool is a subclass of int — must be checked first.
                 if isinstance(value, bool):
                     value = Variant(value, UA_VT.Boolean)
                 elif isinstance(value, float):
                     value = Variant(value, UA_VT.Float)
                 elif isinstance(value, int):
-                    value = Variant(value, UA_VT.Int32)
+                    value = Variant(value, UA_VT.Int16)
                 await self._ctx.address_space.set_value(meta.node, value)
             except Exception as exc:
                 logger.debug("cobot.aas_mirror.error", path=path, error=str(exc))
